move dir-related stuff into dir.rs
