+++ src/ufs/mod.rs
@@ -16,6 +16,7 @@ const MAX_CACHE: Duration = Duration::MAX;
 mod xattr;
 mod symlink;
 mod inode;
+mod dir;
 
 use crate::{
 	blockreader::BlockReader,
@@ -61,26 +62,6 @@ impl Ufs {
 		Ok(Self { file, superblock })
 	}
 
-	fn readdir<T>(
-		&mut self,
-		inr: u64,
-		ino: &Inode,
-		mut f: impl FnMut(&OsStr, InodeNum, FileType) -> Option<T>,
-	) -> IoResult<Option<T>> {
-		let mut block = vec![0u8; self.superblock.bsize as usize];
-		let frag = self.superblock.frag as u64;
-
-		for blkidx in 0..(ino.blocks / frag) {
-			let size = self.inode_read_block(inr, ino, blkidx, &mut block)?;
-
-			let x = readdir_block(inr, &block[0..size], self.file.config(), &mut f)?;
-			if x.is_some() {
-				return Ok(x);
-			}
-		}
-		Ok(None)
-	}
-
 }
 
 fn run<T>(f: impl FnOnce() -> IoResult<T>) -> Result<T, c_int> {
@@ -96,59 +77,6 @@ fn transino(ino: u64) -> u64 {
 	} else {
 		ino
 	}
-}
-
-fn readdir_block<T>(
-	inr: u64,
-	block: &[u8],
-	config: Config,
-	mut f: impl FnMut(&OsStr, InodeNum, FileType) -> Option<T>,
-) -> IoResult<Option<T>> {
-	let mut name = [0u8; UFS_MAXNAMELEN + 1];
-	let file = Cursor::new(block);
-	let mut file = Decoder::new(file, config);
-
-	loop {
-		let Ok(ino) = file.decode::<InodeNum>() else {
-			break;
-		};
-		if ino == 0 {
-			break;
-		}
-
-		let reclen: u16 = file.decode()?;
-		let kind: u8 = file.decode()?;
-		let namelen: u8 = file.decode()?;
-		let name = &mut name[0..namelen.into()];
-		file.read(name)?;
-
-		// skip remaining bytes of record, if any
-		let off = reclen - (namelen as u16) - 8;
-		file.seek_relative(off as i64)?;
-
-		let name = unsafe { OsStr::from_encoded_bytes_unchecked(name) };
-		let kind = match kind {
-			DT_FIFO => FileType::NamedPipe,
-			DT_CHR => FileType::CharDevice,
-			DT_DIR => FileType::Directory,
-			DT_BLK => FileType::BlockDevice,
-			DT_REG => FileType::RegularFile,
-			DT_LNK => FileType::Symlink,
-			DT_SOCK => FileType::Socket,
-			DT_WHT => {
-				log::warn!("readdir_block({inr}): encountered a whiteout entry: {name:?}");
-				continue;
-			}
-			DT_UNKNOWN => todo!("DT_UNKNOWN: {ino}"),
-			_ => panic!("invalid filetype: {kind}"),
-		};
-		let res = f(name, ino, kind);
-		if res.is_some() {
-			return Ok(res);
-		}
-	}
-
-	Ok(None)
 }
 
 impl Filesystem for Ufs {
@@ -225,11 +153,9 @@ impl Filesystem for Ufs {
 				return Ok(());
 			}
 
-			let ino = self.read_inode(inr)?;
-
 			let mut i = 0;
 
-			self.readdir(inr, &ino, |name, inr, kind| {
+			self.dir_iter(inr, |name, inr, kind| {
 				i += 1;
 				if i > offset && reply.add(inr.into(), i, kind, name) {
 					return Some(());
@@ -249,24 +175,9 @@ impl Filesystem for Ufs {
 		let pinr = transino(pinr);
 
 		let mut f = || {
-			let pino = self.read_inode(pinr)?;
-
-			let x = self.readdir(pinr, &pino, |name2, inr, _| {
-				if name == name2 {
-					Some(inr.into())
-				} else {
-					None
-				}
-			});
-
-			match x {
-				Ok(Some(inr)) => {
-					let ino = self.read_inode(inr)?;
-					Ok((ino.as_fileattr(inr), ino.gen))
-				}
-				Ok(None) => Err(err!(ENOENT)),
-				Err(e) => Err(e),
-			}
+			let inr = self.dir_lookup(pinr, name)?;
+			let ino = self.read_inode(inr)?;
+			Ok::<_, IoError>((ino.as_fileattr(inr), ino.gen))
 		};
 
 		match f() {
